requirements.txt
Flask==3.0.2
gunicorn==23.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
Pillow==9.5.0
psycopg2-binary==2.9.9 